Refactor Permission and add tests with repository support
Updated `Permission` entity to use `Long` as `subject` type, added cascade deletion rules, and enabled auditing. Introduced `PermissionRepositoryPort` with comprehensive tests to verify CRUD, referential integrity, concurrency, and audit functionality.

+++ src/main/java/com/tritech/tricore/core/domain/primarykeys/PermissionId.java
@@ -27,7 +27,7 @@ public class PermissionId implements Serializable {
     /**
      * The unique identifier of the user associated with the permission.
      */
-    private Integer subject;
+    private Long subject;
     /**
      * The name of the group associated with the permission.
      */
